ecmcEcSlave: Add status word
Bunch status infor for slave into one status word:
- bit 0      : online
- bit 1      : operational
- bit 2      : alstate.init
- bit 3      : alstate.preop
- bit 4      : alstate.safeop
- bit5       : alstate.op
- bit 16..32 : entrycounter

Less asyn params needed (before 4 now 1).

+++ ecmcApp/src/ethercat/ecmcEcSlave.h
@@ -124,7 +124,7 @@ class ecmcEcSlave : public ecmcError {
   ec_slave_config_t *slaveConfig_;
   ecmcEcSyncManager *syncManagerArray_[EC_MAX_SYNC_MANAGERS];
   ecmcEcEntry *entryList_[EC_MAX_ENTRIES];
-  int16_t entryCounter_;
+  uint32_t entryCounter_;
   int pdosArrayIndex_;
   int syncManArrayIndex_;
   int syncManCounter_;
